# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- SPECIALTIES ----
  /specialties:
    get:
      operationId: listSpecialties
      tags: [specialties]
      summary: List all specialties
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of specialties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialty"
    post:
      operationId: createSpecialty
      tags: [specialties]
      summary: Create specialty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
  /specialties/{id}:
    delete:
      operationId: deleteSpecialty
      tags: [specialties]
      summary: Delete specialty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- LOCATIONS ----
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- DOCTORS ----
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        doctorCount:
          type: integer
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string

    # SPECIALTIES
    Specialty:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        departmentId:
          type: ["integer", "null"]
    SpecialtyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        departmentId:
          type: integer

    # LOCATIONS
    Location:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        district:
          type: ["string", "null"]
    LocationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        district:
          type: string

    # DOCTORS
...